Promote plugin system from experimental to GA
Move the plugin system to a stable public surface at `prefect.plugins`,
backed by private implementation under `prefect._internal.plugins.*`.
Settings move to `settings.plugins.*` / `PREFECT_PLUGINS_*`. The diagnose
command graduates to `prefect plugins diagnose`. Existing users on the
experimental paths continue to work — old imports, env vars, settings
attributes, and CLI commands resolve to the new locations and emit
DeprecationWarnings pointing at the GA path.

A new general-purpose `deprecated_module_attrs` helper in
`prefect._internal.compatibility` powers the import shims, deliberately
separate from the v2->v3-specific `getattr_migration` machinery.
