+++ docs/v3/advanced/self-hosted.mdx
@@ -116,9 +116,9 @@ To use AWS IAM authentication for your PostgreSQL database (experimental):
 
 2. **Create an IAM policy** with `rds-db:connect` permission and attach it to your IAM user/role.
 
-3. **Enable experimental plugins and IAM authentication**:
+3. **Enable plugins and IAM authentication**:
    ```bash
-   export PREFECT_EXPERIMENTS_PLUGINS_ENABLED=true
+   export PREFECT_PLUGINS_ENABLED=true
    export PREFECT_INTEGRATIONS_AWS_RDS_IAM_ENABLED=true
    # Optional: export PREFECT_INTEGRATIONS_AWS_RDS_IAM_REGION_NAME=us-east-1
    ```
