+++ src/prefect/_experimental/plugins/diagnostics.py
@@ -1,25 +1,9 @@
-"""
-Diagnostic data structures for plugin system.
-"""
+"""Deprecated. Use `prefect.plugins` instead."""
 
-from __future__ import annotations
+from prefect._internal.compatibility.deprecated_paths import deprecated_module_attrs
 
-from dataclasses import dataclass
-
-
-@dataclass
-class SetupSummary:
-    """
-    Summary of a plugin's setup_environment execution.
-
-    Attributes:
-        plugin: Name of the plugin
-        env_preview: Preview of environment variables set (with redacted values)
-        note: Human-readable note from the plugin, if any
-        error: Error message if the plugin failed, or None if successful
-    """
-
-    plugin: str
-    env_preview: dict[str, str]
-    note: str | None
-    error: str | None
+__getattr__ = deprecated_module_attrs(
+    __name__,
+    "prefect.plugins",
+    ("SetupSummary",),
+)
