+++ src/prefect/_internal/plugins/__init__.py
@@ -1,0 +1,7 @@
+"""
+Internal implementation of Prefect's plugin system.
+
+Public API for these modules is re-exported from `prefect.plugins`. Importing
+directly from `prefect._internal.plugins.*` is not part of the supported
+contract and may change without notice.
+"""
