+++ src/prefect/plugins.py
@@ -1,74 +1,37 @@
 """
-Utilities for loading plugins that extend Prefect's functionality.
+Prefect plugin system.
 
-Plugins are detected by entry point definitions in package setup files.
+Provides startup hooks that allow third-party packages to run initialization
+code (e.g., set environment variables, configure DB connections) before
+Prefect CLI commands, workers, or scripts begin work, plus loaders for the
+`prefect.collections` entry point group used by Prefect collection packages.
 
-Currently supported entrypoints:
-    - prefect.collections: Identifies this package as a Prefect collection that
-        should be imported when Prefect is imported.
+This module is the stable public surface for the plugin system. Implementation
+modules live under `prefect._internal.plugins.*` and may change without notice.
 """
 
-from importlib.metadata import EntryPoints, entry_points
-from types import ModuleType
-from typing import Any, Union
+from prefect._internal.plugins.collections import (
+    load_prefect_collections,
+    safe_load_entrypoints,
+)
+from prefect._internal.plugins.diagnostics import SetupSummary
+from prefect._internal.plugins.manager import register_hook
+from prefect._internal.plugins.spec import (
+    PREFECT_PLUGIN_API_VERSION,
+    HookContext,
+    HookSpec,
+    SetupResult,
+)
+from prefect._internal.plugins.startup import run_startup_hooks
 
-import prefect.settings
-
-_collections: Union[None, dict[str, Union[ModuleType, Exception]]] = None
-
-
-def safe_load_entrypoints(entrypoints: EntryPoints) -> dict[str, Union[Exception, Any]]:
-    """
-    Load entry points for a group capturing any exceptions that occur.
-    """
-    # TODO: `load()` claims to return module types but could return arbitrary types
-    #       too. We can cast the return type if we want to be more correct. We may
-    #       also want to validate the type for the group for entrypoints that have
-    #       a specific type we expect.
-
-    results: dict[str, Union[Exception, Any]] = {}
-
-    for entrypoint in entrypoints:
-        result = None
-        try:
-            result = entrypoint.load()
-        except Exception as exc:
-            result = exc
-
-        results[entrypoint.name or entrypoint.value] = result
-
-    return results
-
-
-def load_prefect_collections() -> dict[str, Union[ModuleType, Exception]]:
-    """
-    Load all Prefect collections that define an entrypoint in the group
-    `prefect.collections`.
-    """
-    global _collections
-
-    if _collections is not None:
-        return _collections
-
-    collection_entrypoints: EntryPoints = entry_points(group="prefect.collections")
-    collections: dict[str, Union[Exception, Any]] = safe_load_entrypoints(
-        collection_entrypoints
-    )
-
-    # TODO: Consider the utility of this once we've established this pattern.
-    #       We cannot use a logger here because logging is not yet initialized.
-    #       It would be nice if logging was initialized so we could log failures
-    #       at least.
-    for name, result in collections.items():
-        if isinstance(result, Exception):
-            print(
-                # TODO: Use exc_info if we have a logger
-                f"Warning!  Failed to load collection {name!r}:"
-                f" {type(result).__name__}: {result}"
-            )
-        else:
-            if prefect.settings.PREFECT_DEBUG_MODE:
-                print(f"Loaded collection {name!r}.")
-
-    _collections = collections
-    return collections
+__all__ = [
+    "load_prefect_collections",
+    "safe_load_entrypoints",
+    "run_startup_hooks",
+    "HookContext",
+    "SetupResult",
+    "HookSpec",
+    "SetupSummary",
+    "PREFECT_PLUGIN_API_VERSION",
+    "register_hook",
+]
