+++ tests/test_settings.py
@@ -264,15 +264,32 @@ SUPPORTED_SETTINGS = {
     "PREFECT_EXPERIMENTS_PLUGINS_ALLOW": {
         "test_value": "plugin1,plugin2",
         "expected_value": {"plugin1", "plugin2"},
+        "legacy": True,
     },
     "PREFECT_EXPERIMENTS_PLUGINS_DENY": {
         "test_value": "plugin3",
         "expected_value": {"plugin3"},
+        "legacy": True,
     },
-    "PREFECT_EXPERIMENTS_PLUGINS_ENABLED": {"test_value": True},
-    "PREFECT_EXPERIMENTS_PLUGINS_SAFE_MODE": {"test_value": True},
-    "PREFECT_EXPERIMENTS_PLUGINS_SETUP_TIMEOUT_SECONDS": {"test_value": 30.0},
-    "PREFECT_EXPERIMENTS_PLUGINS_STRICT": {"test_value": True},
+    "PREFECT_EXPERIMENTS_PLUGINS_ENABLED": {"test_value": True, "legacy": True},
+    "PREFECT_EXPERIMENTS_PLUGINS_SAFE_MODE": {"test_value": True, "legacy": True},
+    "PREFECT_EXPERIMENTS_PLUGINS_SETUP_TIMEOUT_SECONDS": {
+        "test_value": 30.0,
+        "legacy": True,
+    },
+    "PREFECT_EXPERIMENTS_PLUGINS_STRICT": {"test_value": True, "legacy": True},
+    "PREFECT_PLUGINS_ALLOW": {
+        "test_value": "plugin1,plugin2",
+        "expected_value": {"plugin1", "plugin2"},
+    },
+    "PREFECT_PLUGINS_DENY": {
+        "test_value": "plugin3",
+        "expected_value": {"plugin3"},
+    },
+    "PREFECT_PLUGINS_ENABLED": {"test_value": True},
+    "PREFECT_PLUGINS_SAFE_MODE": {"test_value": True},
+    "PREFECT_PLUGINS_SETUP_TIMEOUT_SECONDS": {"test_value": 30.0},
+    "PREFECT_PLUGINS_STRICT": {"test_value": True},
     "PREFECT_FLOW_DEFAULT_RETRIES": {"test_value": 10, "legacy": True},
     "PREFECT_FLOWS_DEFAULT_RETRIES": {"test_value": 10},
     "PREFECT_FLOW_DEFAULT_RETRY_DELAY_SECONDS": {"test_value": 10, "legacy": True},
